+++ Assets/Android Build Pipeline/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 1.0.6
+- Add new popup field of **'Create Symbols'** that will create `symbols.zip` on build, using for upload to Google Play Console.
+- Remove usage of `BUILD_PIPELINE_ANDROID_ENABLE_BUNDLE_VERSION_CONFIG` and always enable app bundle version field.
+- Add label for display next app bundle version when enabling **'Increment Bundle Version'**.
+- **'Build App Bundle'** and **'Split Application Binary'** will use value from Unity Editor.
+
 ## 1.0.5
 - Require minimal version of Unity Editor 2021.3 LTS. Older of Unity Editor version are no longer support.
 - Change UI.
@@ -17,10 +23,10 @@
     - Move 'open settings' to 'Settings/Android'.
 
 ## 1.0.2
-- Add the option of 'Use Keystore'
+- Add the option of **'Use Keystore'**
 
 ## 1.0.1
-- Add the option of 'Increment Bundle Version'
+- Add the option of **'Increment Bundle Version'**
 
 ## 1.0.0
 - Create the plugins.