+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 
-## v1.0.5 November 29th 2022 - Latest
+## v1.0.6 December 12th 2022 - Latest
+
+`https://github.com/LuviKunG/BuildPipelineAndroid.git#1.0.6`
+
+- Add new popup field of **'Create Symbols'** that will create `symbols.zip` on build, using for upload to Google Play Console.
+- Remove usage of `BUILD_PIPELINE_ANDROID_ENABLE_BUNDLE_VERSION_CONFIG` and always enable app bundle version field.
+- Add label for display next app bundle version when enabling **'Increment Bundle Version'**.
+- **'Build App Bundle'** and **'Split Application Binary'** will use value from Unity Editor.
+
+## v1.0.5 November 29th 2022
 
 `https://github.com/LuviKunG/BuildPipelineAndroid.git#1.0.5`
 
@@ -29,13 +38,13 @@
 
 `https://github.com/LuviKunG/BuildPipelineAndroid.git#1.0.2`
 
-- Add the option of 'Use Keystore'
+- Add the option of **'Use Keystore'**
 
 ## v1.0.1 August 30th 2019
 
 `https://github.com/LuviKunG/BuildPipelineAndroid.git#1.0.1`
 
-- Add the option of 'Increment Bundle Version'
+- Add the option of **'Increment Bundle Version'**
 
 ## v1.0.0 August 30th 2019